Bug#42392: Falcon bugs in recovery of create tablespace Problem: during recovery of workload that does "create tablespace", falcon  would assert with the message "page 1/1 wrong page type, expected 8 got N (where N != 8)
Analysis:
Reason for the error is recent fix for Bug42060, that introduces 
logging for page inventory pages. PIP page #1 needs special 
handling that is missing in the above mentioned bug fix. Unlike 
other PIP pages, this one is created during "create tablespace" 
and should already have first 4 pages marked as allocated
(for special falcon pages header, pip, section and index roots 
that have numbers 1-3). Recovery did not mark those pages as
allocated and they were taken away by some other page types.

Fix:
Don't log PIP if its page number is 1. Instead, recreate PIP and
other special pages during recovery of "create tablespace".


This patch also contains fix for Bug#42392 - removing 
unnecessary TableSpace::open() from TableSpaceManager::bootstrap()

+++ storage/falcon/Dbb.cpp
@@ -149,7 +149,7 @@ Dbb::~Dbb()
 		dbb->close();
 }
 
-Cache* Dbb::create(const char * fileName, int pageSz, int64 cacheSize, FileType fileType, TransId transId, const char *logRoot)
+Cache* Dbb::create(const char * fileName, int pageSz, int64 cacheSize, FileType fileType, TransId transId, const char *logRoot, bool useExistingFile)
 {
 	serialLog = database->serialLog;
 	odsVersion = ODS_VERSION;
@@ -157,7 +157,11 @@ Cache* Dbb::create(const char * fileName, int pageSz, int64 cacheSize, FileType 
 	sequence = 1;
 
 	init(pageSz, (int) ((cacheSize + pageSz - 1) / pageSz));
-	createFile(fileName);
+	if(useExistingFile)
+		openFile(fileName, false);
+	else
+		createFile(fileName);
+
 	try
 		{
 		Hdr::create(this, fileType, transId, logRoot);
@@ -168,7 +172,8 @@ Cache* Dbb::create(const char * fileName, int pageSz, int64 cacheSize, FileType 
 	catch(...)
 		{
 		closeFile();
-		deleteFile();
+		if(!useExistingFile)
+			deleteFile();
 		throw;
 		}
